refactor(#11) : interface dir 생성, 상수는 constant에 몰아주
