+++ src/constants/KeywordList.ts
@@ -1,10 +1,10 @@
 export const keywordList = [
-  '계획 마스터',
+  '🔥계획 마스터',
   '불타는 열정왕',
   '전공스킬 넘사벽',
-  '소통의 귀재',
-  '논리의 왕',
-  '분위기 메이커',
+  '💬소통의 귀재',
+  '🧠논리의 왕',
+  '😎분위기 메이커',
   '꽉찬 리더십',
   '경험가득 고인물',
   '올라운더',
